chore(#36): moves config and server to root pkg
