+++ pkg/config/config_loader_test.go
@@ -1,7 +1,8 @@
 package config_test
 
 import (
-	"github.com/arquillian/ike-prow-plugins/pkg/plugin/config"
+
+	"github.com/arquillian/ike-prow-plugins/pkg/config"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
