+++ pkg/config/config_suite_test.go
@@ -1,4 +1,4 @@
-package config_test
+package config
 
 import (
 	"testing"
