+++ pkg/plugin/test-keeper/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/arquillian/ike-prow-plugins/pkg/plugin/server"
+	"github.com/arquillian/ike-prow-plugins/pkg/server"
 	"github.com/arquillian/ike-prow-plugins/pkg/plugin/test-keeper/plugin"
 	gogh "github.com/google/go-github/github"
 	"k8s.io/test-infra/prow/pluginhelp"
